Move WallElement into its own header/source files
