+++ src/openrct2-ui/input/Shortcuts.cpp
@@ -45,6 +45,7 @@
 #include <openrct2/world/Park.h>
 #include <openrct2/world/Scenery.h>
 #include <openrct2/world/TileInspector.h>
+#include <openrct2/world/tile_element/WallElement.h>
 
 using namespace OpenRCT2;
 using namespace OpenRCT2::Ui;
