+++ src/openrct2-ui/interface/ViewportInteraction.cpp
@@ -52,8 +52,8 @@
 #include <openrct2/world/Park.h>
 #include <openrct2/world/Scenery.h>
 #include <openrct2/world/Surface.h>
-#include <openrct2/world/Wall.h>
 #include <openrct2/world/tile_element/EntranceElement.h>
+#include <openrct2/world/tile_element/WallElement.h>
 
 using namespace OpenRCT2;
 using namespace OpenRCT2::Ui::Windows;
