+++ src/openrct2-ui/windows/Scenery.cpp
@@ -50,6 +50,7 @@
 #include <openrct2/sprites.h>
 #include <openrct2/world/Park.h>
 #include <openrct2/world/Scenery.h>
+#include <openrct2/world/tile_element/WallElement.h>
 
 namespace OpenRCT2::Ui::Windows
 {
