+++ src/openrct2-ui/windows/Sign.cpp
@@ -25,7 +25,7 @@
 #include <openrct2/sprites.h>
 #include <openrct2/world/Banner.h>
 #include <openrct2/world/Scenery.h>
-#include <openrct2/world/Wall.h>
+#include <openrct2/world/tile_element/WallElement.h>
 
 namespace OpenRCT2::Ui::Windows
 {
