+++ src/openrct2/actions/FootpathAdditionRemoveAction.cpp
@@ -20,7 +20,6 @@
 #include "../world/Footpath.h"
 #include "../world/Location.hpp"
 #include "../world/Park.h"
-#include "../world/Wall.h"
 
 using namespace OpenRCT2;
