+++ src/openrct2/ride/TrackDesign.cpp
@@ -57,7 +57,6 @@
 #include "../world/Park.h"
 #include "../world/Scenery.h"
 #include "../world/Surface.h"
-#include "../world/Wall.h"
 #include "../world/tile_element/EntranceElement.h"
 #include "../world/tile_element/Slope.h"
 #include "Ride.h"
