+++ src/openrct2/ride/TrackDesignSave.cpp
@@ -25,8 +25,8 @@
 #include "../windows/Intent.h"
 #include "../world/Footpath.h"
 #include "../world/Scenery.h"
-#include "../world/Wall.h"
 #include "../world/tile_element/EntranceElement.h"
+#include "../world/tile_element/WallElement.h"
 #include "RideData.h"
 #include "Station.h"
 #include "Track.h"
