+++ src/openrct2/world/ConstructionClearance.cpp
@@ -24,6 +24,7 @@
 #include "Surface.h"
 #include "tile_element/EntranceElement.h"
 #include "tile_element/Slope.h"
+#include "tile_element/WallElement.h"
 
 using namespace OpenRCT2;
