+++ src/openrct2/world/Wall.cpp
@@ -9,23 +9,8 @@
 
 #include "Wall.h"
 
-#include "../Cheats.h"
-#include "../Game.h"
-#include "../OpenRCT2.h"
-#include "../localisation/StringIds.h"
-#include "../management/Finance.h"
-#include "../network/network.h"
-#include "../object/ObjectEntryManager.h"
-#include "../object/WallSceneryEntry.h"
-#include "../ride/Track.h"
-#include "../ride/TrackData.h"
-#include "Banner.h"
 #include "Map.h"
-#include "MapAnimation.h"
-#include "Park.h"
-#include "Scenery.h"
-#include "Surface.h"
-#include "Wall.h"
+#include "tile_element/WallElement.h"
 
 /**
  *
@@ -77,112 +62,6 @@ void WallRemoveIntersectingWalls(const CoordsXYRangedZ& wallPos, Direction direc
         TileElementRemove(tileElement);
         tileElement--;
     } while (!(tileElement++)->IsLastForTile());
-}
-
-uint8_t WallElement::GetSlope() const
-{
-    return (Type & kTileElementQuadrantMask) >> 6;
-}
-
-void WallElement::SetSlope(uint8_t newSlope)
-{
-    Type &= ~kTileElementQuadrantMask;
-    Type |= (newSlope << 6);
-}
-
-colour_t WallElement::GetPrimaryColour() const
-{
-    return colour_1;
-}
-
-colour_t WallElement::GetSecondaryColour() const
-{
-    return colour_2;
-}
-
-colour_t WallElement::GetTertiaryColour() const
-{
-    return colour_3;
-}
-
-void WallElement::SetPrimaryColour(colour_t newColour)
-{
-    colour_1 = newColour;
-}
-
-void WallElement::SetSecondaryColour(colour_t newColour)
-{
-    colour_2 = newColour;
-}
-
-void WallElement::SetTertiaryColour(colour_t newColour)
-{
-    colour_3 = newColour;
-}
-
-uint8_t WallElement::GetAnimationFrame() const
-{
-    return (animation >> 3) & 0xF;
-}
-
-void WallElement::SetAnimationFrame(uint8_t frameNum)
-{
-    animation &= WALL_ANIMATION_FLAG_ALL_FLAGS;
-    animation |= (frameNum & 0xF) << 3;
-}
-
-uint16_t WallElement::GetEntryIndex() const
-{
-    return entryIndex;
-}
-
-const WallSceneryEntry* WallElement::GetEntry() const
-{
-    return OpenRCT2::ObjectManager::GetObjectEntry<WallSceneryEntry>(entryIndex);
-}
-
-void WallElement::SetEntryIndex(uint16_t newIndex)
-{
-    entryIndex = newIndex;
-}
-
-Banner* WallElement::GetBanner() const
-{
-    return ::GetBanner(GetBannerIndex());
-}
-
-BannerIndex WallElement::GetBannerIndex() const
-{
-    return banner_index;
-}
-
-void WallElement::SetBannerIndex(BannerIndex newIndex)
-{
-    banner_index = newIndex;
-}
-
-bool WallElement::IsAcrossTrack() const
-{
-    return (animation & WALL_ANIMATION_FLAG_ACROSS_TRACK) != 0;
-}
-
-void WallElement::SetAcrossTrack(bool acrossTrack)
-{
-    animation &= ~WALL_ANIMATION_FLAG_ACROSS_TRACK;
-    if (acrossTrack)
-        animation |= WALL_ANIMATION_FLAG_ACROSS_TRACK;
-}
-
-bool WallElement::AnimationIsBackwards() const
-{
-    return (animation & WALL_ANIMATION_FLAG_DIRECTION_BACKWARD) != 0;
-}
-
-void WallElement::SetAnimationIsBackwards(bool isBackwards)
-{
-    animation &= ~WALL_ANIMATION_FLAG_DIRECTION_BACKWARD;
-    if (isBackwards)
-        animation |= WALL_ANIMATION_FLAG_DIRECTION_BACKWARD;
 }
 
 #pragma region Edge Slopes Table
