+++ src/openrct2/world/Wall.h
@@ -11,14 +11,6 @@
 
 #include "TileElement.h"
 
-enum
-{
-    WALL_ANIMATION_FLAG_ACROSS_TRACK = (1 << 2),
-    // 3 - 6 animation frame number
-    WALL_ANIMATION_FLAG_DIRECTION_BACKWARD = (1 << 7),
-    WALL_ANIMATION_FLAG_ALL_FLAGS = WALL_ANIMATION_FLAG_ACROSS_TRACK | WALL_ANIMATION_FLAG_DIRECTION_BACKWARD
-};
-
 enum EDGE_SLOPE
 {
     EDGE_SLOPE_UPWARDS = (1 << 0),
