+++ test/tests/TileElementsView.cpp
@@ -19,6 +19,7 @@
 #include <openrct2/world/Map.h>
 #include <openrct2/world/TileElementsView.h>
 #include <openrct2/world/tile_element/EntranceElement.h>
+#include <openrct2/world/tile_element/WallElement.h>
 
 using namespace OpenRCT2;
